Move extra functions out of push.js
