+++ src/commands/listCommits.js
@@ -1,0 +1,49 @@
+import path from 'path'
+
+import { GitObjectManager, GitRefManager } from '../managers'
+import { FileSystem, GitCommit } from '../models'
+
+export async function listCommits ({
+  dir,
+  gitdir = path.join(dir, '.git'),
+  fs: _fs,
+  start,
+  finish
+}) {
+  const fs = new FileSystem(_fs)
+  let startingSet = new Set()
+  let finishingSet = new Set()
+  for (let ref of start) {
+    startingSet.add(await GitRefManager.resolve({ fs, gitdir, ref }))
+  }
+  for (let ref of finish) {
+    // We may not have these refs locally so we must try/catch
+    try {
+      let oid = await GitRefManager.resolve({ fs, gitdir, ref })
+      finishingSet.add(oid)
+    } catch (err) {}
+  }
+  let visited = new Set()
+  // Because git commits are named by their hash, there is no
+  // way to construct a cycle. Therefore we won't worry about
+  // setting a default recursion limit.
+  async function walk (oid) {
+    visited.add(oid)
+    let { type, object } = await GitObjectManager.read({ fs, gitdir, oid })
+    if (type !== 'commit') {
+      throw new Error(`Expected type commit but type is ${type}`)
+    }
+    let commit = GitCommit.from(object)
+    let parents = commit.headers().parent
+    for (oid of parents) {
+      if (!finishingSet.has(oid) && !visited.has(oid)) {
+        await walk(oid)
+      }
+    }
+  }
+  // Let's go walking!
+  for (let oid of startingSet) {
+    await walk(oid)
+  }
+  return visited
+}
