+++ src/commands/types.js
@@ -1,0 +1,8 @@
+export const types = {
+  commit: 0b0010000,
+  tree: 0b0100000,
+  blob: 0b0110000,
+  tag: 0b1000000,
+  ofs_delta: 0b1100000,
+  ref_delta: 0b1110000
+}
