+++ src/internal-apis.js
@@ -5,6 +5,6 @@ import * as utils from './utils'
 export { managers }
 export { models }
 export { utils }
-export { listCommits } from './commands/push'
-export { listObjects } from './commands/push'
-export { pack } from './commands/push'
+export { listCommits } from './commands/listCommits'
+export { listObjects } from './commands/listObjects'
+export { pack } from './commands/pack'
